new release_04.02.00

+++ frontend/src/container/ReduxTlmDispatcher.jsx
@@ -13,6 +13,7 @@ import {
   newFlashMessage,
   addNotification,
   addWorkspaceContentList,
+  addWorkspaceList,
   addWorkspaceMember,
   deleteWorkspaceContentList,
   removeWorkspaceMember,
@@ -108,7 +109,7 @@ export class ReduxTlmDispatcher extends React.Component {
       this.props.user.userId !== data.fields.author.user_id &&
       !EXCLUDED_NOTIFICATION_TYPE_PREFIXES.some(type => data.event_type.startsWith(type))
     ) {
-      this.props.dispatch(addNotification(data))
+      this.props.dispatch(addNotification(data, this.props.workspaceList))
     }
   }
 
@@ -140,15 +141,25 @@ export class ReduxTlmDispatcher extends React.Component {
 
   handleMemberCreated = data => {
     const { props } = this
-    props.dispatch(addWorkspaceMember(data.fields.user, data.fields.workspace.workspace_id, data.fields.member))
-    if (props.user.userId === data.fields.user.user_id) {
-      props.dispatch(removeAccessibleWorkspace(data.fields.workspace))
+    if (props.user.userId === data.fields.user.user_id || props.workspaceList.find(space => space.id === data.fields.workspace.workspace_id)) {
+      // NOTE - RJ & MP - 2022-02-18
+      // When receiving a member created TLM, it is possible that we haven't added the workspace itself yet
+      // In this case, addWorkspaceMember does nothing.
+      // We actually noticed that the member created TLM arrives before the workspace created TLM.
+      // Let's add the workpace first to avoid this.
+      props.dispatch(addWorkspaceList([data.fields.workspace]))
+
+      props.dispatch(addWorkspaceMember(data.fields.user, data.fields.workspace.workspace_id, data.fields.member))
+      if (props.user.userId === data.fields.user.user_id) {
+        props.dispatch(removeAccessibleWorkspace(data.fields.workspace))
+      }
+      this.handleNotification(data)
     }
-    this.handleNotification(data)
   }
 
   handleMemberModified = data => {
     const { props } = this
+    props.dispatch(addWorkspaceList([data.fields.workspace]))
     props.dispatch(updateWorkspaceMember(data.fields.user, data.fields.workspace.workspace_id, data.fields.member))
     this.handleNotification(data)
   }
@@ -247,7 +258,7 @@ export class ReduxTlmDispatcher extends React.Component {
   }
 
   handleUserChanged = data => {
-    this.props.dispatch(addNotification(data))
+    this.props.dispatch(addNotification(data, this.props.workspaceList))
   }
 
   handleWorkspaceSubscriptionCreated = data => {
@@ -289,5 +300,5 @@ export class ReduxTlmDispatcher extends React.Component {
   }
 }
 
-const mapStateToProps = ({ workspaceContentList, user }) => ({ workspaceContentList, user })
+const mapStateToProps = ({ workspaceList, workspaceContentList, user }) => ({ workspaceList, workspaceContentList, user })
 export default translate()(connect(mapStateToProps)(TracimComponent(ReduxTlmDispatcher)))
